+++ angular.json
@@ -189,7 +189,7 @@
     },
     "@flosportsinc/ng-media-source-extensions": {
       "root": "projects/flosportsinc/ng-media-source-extensions",
-      "sourceRoot": "projects/flosportsinc/ng-media-source-extensions/src",
+      "sourceRoot": "projects/flosportsinc/ng-media-source-extensions",
       "projectType": "library",
       "prefix": "flo",
       "schematics": {
@@ -210,7 +210,7 @@
           "builder": "@angular-devkit/build-angular:karma",
           "options": {
             "codeCoverage": true,
-            "main": "projects/flosportsinc/ng-media-source-extensions/src/test.ts",
+            "main": "projects/flosportsinc/ng-media-source-extensions/test.ts",
             "tsConfig": "projects/flosportsinc/ng-media-source-extensions/tsconfig.spec.json",
             "karmaConfig": "projects/flosportsinc/ng-media-source-extensions/karma.conf.js"
           }
@@ -261,86 +261,6 @@
             "tsConfig": [
               "projects/flosportsinc/ng-universal-services/tsconfig.lib.json",
               "projects/flosportsinc/ng-universal-services/tsconfig.spec.json"
-            ],
-            "exclude": ["**/node_modules/**"]
-          }
-        }
-      }
-    },
-    "@flosportsinc/ng-media-source-extensions-dash": {
-      "root": "projects/flosportsinc/ng-media-source-extensions-dash",
-      "sourceRoot": "projects/flosportsinc/ng-media-source-extensions-dash/src",
-      "projectType": "library",
-      "prefix": "flo",
-      "schematics": {
-        "@schematics/angular:component": {
-          "styleext": "scss",
-          "changeDetection": "OnPush"
-        }
-      },
-      "architect": {
-        "build": {
-          "builder": "@angular-devkit/build-ng-packagr:build",
-          "options": {
-            "tsConfig": "projects/flosportsinc/ng-media-source-extensions-dash/tsconfig.lib.json",
-            "project": "projects/flosportsinc/ng-media-source-extensions-dash/ng-package.json"
-          }
-        },
-        "test": {
-          "builder": "@angular-devkit/build-angular:karma",
-          "options": {
-            "codeCoverage": true,
-            "main": "projects/flosportsinc/ng-media-source-extensions-dash/src/test.ts",
-            "tsConfig": "projects/flosportsinc/ng-media-source-extensions-dash/tsconfig.spec.json",
-            "karmaConfig": "projects/flosportsinc/ng-media-source-extensions-dash/karma.conf.js"
-          }
-        },
-        "lint": {
-          "builder": "@angular-devkit/build-angular:tslint",
-          "options": {
-            "tsConfig": [
-              "projects/flosportsinc/ng-media-source-extensions-dash/tsconfig.lib.json",
-              "projects/flosportsinc/ng-media-source-extensions-dash/tsconfig.spec.json"
-            ],
-            "exclude": ["**/node_modules/**"]
-          }
-        }
-      }
-    },
-    "@flosportsinc/ng-media-source-extensions-hls": {
-      "root": "projects/flosportsinc/ng-media-source-extensions-hls",
-      "sourceRoot": "projects/flosportsinc/ng-media-source-extensions-hls/src",
-      "projectType": "library",
-      "prefix": "flo",
-      "schematics": {
-        "@schematics/angular:component": {
-          "styleext": "scss",
-          "changeDetection": "OnPush"
-        }
-      },
-      "architect": {
-        "build": {
-          "builder": "@angular-devkit/build-ng-packagr:build",
-          "options": {
-            "tsConfig": "projects/flosportsinc/ng-media-source-extensions-hls/tsconfig.lib.json",
-            "project": "projects/flosportsinc/ng-media-source-extensions-hls/ng-package.json"
-          }
-        },
-        "test": {
-          "builder": "@angular-devkit/build-angular:karma",
-          "options": {
-            "codeCoverage": true,
-            "main": "projects/flosportsinc/ng-media-source-extensions-hls/src/test.ts",
-            "tsConfig": "projects/flosportsinc/ng-media-source-extensions-hls/tsconfig.spec.json",
-            "karmaConfig": "projects/flosportsinc/ng-media-source-extensions-hls/karma.conf.js"
-          }
-        },
-        "lint": {
-          "builder": "@angular-devkit/build-angular:tslint",
-          "options": {
-            "tsConfig": [
-              "projects/flosportsinc/ng-media-source-extensions-hls/tsconfig.lib.json",
-              "projects/flosportsinc/ng-media-source-extensions-hls/tsconfig.spec.json"
             ],
             "exclude": ["**/node_modules/**"]
           }
