+++ package.json
@@ -53,7 +53,7 @@
     "@angular/platform-browser": "^8.0.0-beta.2",
     "@angular/platform-browser-dynamic": "^8.0.0-beta.2",
     "@angular/platform-server": "^8.0.0-beta.2",
-    "@angular/pwa": "^0.12.4",
+    "@angular/pwa": "^0.13.0",
     "@angular/router": "^8.0.0-beta.2",
     "@angular/service-worker": "^8.0.0-beta.2",
     "@nguniversal/common": "^7.1.0",
@@ -75,8 +75,8 @@
     "zone.js": "^0.8.29"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.12.4",
-    "@angular-devkit/build-ng-packagr": "^0.12.4",
+    "@angular-devkit/build-angular": "^0.13.0",
+    "@angular-devkit/build-ng-packagr": "^0.13.0",
     "@angular/cli": "^7.3.0",
     "@angular/compiler-cli": "^8.0.0-beta.2",
     "@fortawesome/fontawesome-free": "^5.7.1",
