--- projects/flosportsinc/ng-media-source-extensions-dash/README.md
@@ -1,6 +1,0 @@
-# FloSports ng-media-source-extensions-dash
-For cross platform Dynamic Adaptive Streaming over HTTP (DASH) support in Angular.
-
-```sh
-npm i @flosportsinc/ng-media-source-extensions-dash
-```
