--- projects/flosportsinc/ng-media-source-extensions-dash/karma.conf.js
@@ -1,2 +1,0 @@
-const config = require('../../../src/karma.conf.shared')
-module.exports = config('../coverage/ng-media-source-extensions-dash')
