--- projects/flosportsinc/ng-media-source-extensions-dash/ng-package.json
@@ -1,15 +1,0 @@
-{
-  "$schema": "../../../node_modules/ng-packagr/ng-package.schema.json",
-  "dest": "../../../dist/flosportsinc/ng-media-source-extensions-dash",
-  "lib": {
-    "entryFile": "src/public_api.ts",
-    "umdModuleIds": {
-      "dashjs": "dashjs",
-      "typescript-monads": "typescriptMonads"
-    }
-  },
-  "whitelistedNonPeerDependencies": [
-    "typescript-monads",
-    "dashjs"
-  ]
-}
