--- projects/flosportsinc/ng-media-source-extensions-dash/package.json
@@ -1,24 +1,0 @@
-{
-  "name": "@flosportsinc/ng-media-source-extensions-dash",
-  "version": "0.0.0-development",
-  "author": "Patrick Michalina <patrickmichalina@mac.com>",
-  "homepage": "https://flosports-component-library.herokuapp.com",
-  "bugs": {
-    "url": "https://github.com/flocasts/flo-angular/issues"
-  },
-  "contributors": [],
-  "license": "MIT",
-  "private": false,
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/flocasts/flo-angular"
-  },
-  "peerDependencies": {
-    "@angular/common": "^7.0.0",
-    "@angular/core": "^7.0.0"
-  },
-  "dependencies": {
-    "dashjs": "2.9.2"
-  },
-  "keywords": ["angular", "universal", "flosports", "dash", "streaming", "live"]
-}
