--- projects/flosportsinc/ng-media-source-extensions-dash/src/public_api.ts
@@ -1,5 +1,0 @@
-/*
- * Public API Surface of ng-media-source-extensions-dash
- */
-
-export * from './lib/dash.module'
