--- projects/flosportsinc/ng-media-source-extensions-dash/tslint.json
@@ -1,15 +1,0 @@
-{
-  "extends": "../../../tslint.json",
-  "rules": {
-    "directive-selector": [
-      true,
-      "attribute",
-      "camelCase"
-    ],
-    "component-selector": [
-      true,
-      "element",
-      "kebab-case"
-    ]
-  }
-}
