--- projects/flosportsinc/ng-media-source-extensions-hls/README.md
@@ -1,6 +1,0 @@
-# FloSports ng-media-source-extensions-hls
-For cross platform HTTP Live Streaming (HLS) support in Angular.
-
-```sh
-npm i @flosportsinc/ng-media-source-extensions-hls
-```
