--- projects/flosportsinc/ng-media-source-extensions-hls/ng-package.json
@@ -1,15 +1,0 @@
-{
-  "$schema": "../../../node_modules/ng-packagr/ng-package.schema.json",
-  "dest": "../../../dist/flosportsinc/ng-media-source-extensions-hls",
-  "lib": {
-    "entryFile": "src/public_api.ts",
-    "umdModuleIds": {
-      "hls.js": "hls.js",
-      "typescript-monads": "typescriptMonads"
-    }
-  },
-  "whitelistedNonPeerDependencies": [
-    "typescript-monads",
-    "hls.js"
-  ]
-}
