--- projects/flosportsinc/ng-media-source-extensions-hls/src/public_api.ts
@@ -1,5 +1,0 @@
-/*
- * Public API Surface of ng-media-source-extensions-hls
- */
-
-export * from './lib/hls.module'
