+++ projects/flosportsinc/ng-media-source-extensions/core/mse.tokens.ts
@@ -1,12 +1,11 @@
-import { InjectionToken } from '@angular/core'
 import { Subject } from 'rxjs'
 
-export const SUPPORTS_TARGET_VIA_MEDIA_SOURCE_EXTENSION = new InjectionToken<IMsePlatformSupportCheck>('flo.mse.isSupportedViaMediaSource')
-export const SUPPORTS_MSE_TARGET_NATIVELY = new InjectionToken<IVideoElementSupportsTargetMseCheckContext>('flo.mse.hasNativeSupport')
-export const MEDIA_SOURCE_EXTENSION_LIBRARY_INIT_TASK = new InjectionToken<IMseInit<any, any>>('flo.mse.lib.init')
-export const MEDIA_SOURCE_EXTENSION_LIBRARY_SRC_CHANGE_TASK = new InjectionToken<IMseSrcChangeFunc<any>>('flo.mse.lib.src.change')
-export const MEDIA_SOURCE_EXTENSION_LIBRARY_DESTROY_TASK = new InjectionToken<IMseDestroy<any>>('flo.mse.lib.create')
-export const MEDIA_SOURCE_EXTENSION_PATTERN_MATCH = new InjectionToken<IMsePatternCheck>('flo.mse.lib.pattern')
+export const SUPPORTS_TARGET_VIA_MEDIA_SOURCE_EXTENSION = 'flo.mse.isSupportedViaMediaSource'
+export const SUPPORTS_MSE_TARGET_NATIVELY = 'flo.mse.hasNativeSupport'
+export const MEDIA_SOURCE_EXTENSION_LIBRARY_INIT_TASK = 'flo.mse.lib.init'
+export const MEDIA_SOURCE_EXTENSION_LIBRARY_SRC_CHANGE_TASK = 'flo.mse.lib.src.change'
+export const MEDIA_SOURCE_EXTENSION_LIBRARY_DESTROY_TASK = 'flo.mse.lib.create'
+export const MEDIA_SOURCE_EXTENSION_PATTERN_MATCH = 'flo.mse.lib.pattern'
 
 export interface IMseExecutionContext<TFunc> {
   readonly exectionKey: string
