+++ projects/flosportsinc/ng-media-source-extensions/dash/mse.directive.ts
@@ -1,0 +1,1 @@
+../core/mse.directive.ts
