+++ projects/flosportsinc/ng-media-source-extensions/dash/mse.module.ts
@@ -1,0 +1,1 @@
+../core/mse.module.ts
