+++ projects/flosportsinc/ng-media-source-extensions/dash/ng-package.json
@@ -1,0 +1,5 @@
+{
+  "lib": {
+    "entryFile": "public_api.ts"
+  }
+}
