+++ projects/flosportsinc/ng-media-source-extensions/hls/hls.module.spec.ts
@@ -6,9 +6,9 @@ import { TestBed, async } from '@angular/core/testing'
 import * as Hls from 'hls.js'
 import {
   IMseSrcChangeOptions, SUPPORTS_TARGET_VIA_MEDIA_SOURCE_EXTENSION,
-  SUPPORTS_MSE_TARGET_NATIVELY,
-  MseDirective
-} from '@flosportsinc/ng-media-source-extensions'
+  SUPPORTS_MSE_TARGET_NATIVELY
+} from './mse.tokens'
+import { MseDirective } from './mse.directive'
 import { SimpleChange, Component, Input, NgModule } from '@angular/core'
 import { take } from 'rxjs/operators'
 import { Subject, ObjectUnsubscribedError } from 'rxjs'
