+++ projects/flosportsinc/ng-media-source-extensions/hls/hls.module.ts
@@ -1,5 +1,4 @@
 import {
-  MseModule,
   SUPPORTS_TARGET_VIA_MEDIA_SOURCE_EXTENSION, IMseDestroyFunc,
   MEDIA_SOURCE_EXTENSION_LIBRARY_DESTROY_TASK,
   IVideoElementSupportsTargetMseCheck,
@@ -15,7 +14,8 @@ import {
   IMsePatternCheckFunc,
   MEDIA_SOURCE_EXTENSION_PATTERN_MATCH,
   IVideoElementSupportsTargetMseCheckContext
-} from '@flosportsinc/ng-media-source-extensions'
+} from './mse.tokens'
+import { MseModule } from './mse.module'
 import { NgModule } from '@angular/core'
 import * as Hls from 'hls.js'
