+++ projects/flosportsinc/ng-media-source-extensions/hls/package.json
@@ -1,0 +1,3 @@
+{
+  "name": "@flosportsinc/ng-media-source-extensions"
+}
