+++ projects/flosportsinc/ng-media-source-extensions/karma.conf.js
@@ -1,2 +1,2 @@
 const config = require('../../../src/karma.conf.shared')
-module.exports = config('../coverage/ng-media-source-extensions')
+module.exports = config('../coverage')
