+++ projects/flosportsinc/ng-media-source-extensions/ng-package.json
@@ -2,7 +2,7 @@
   "$schema": "../../../node_modules/ng-packagr/ng-package.schema.json",
   "dest": "../../../dist/flosportsinc/ng-media-source-extensions",
   "lib": {
-    "entryFile": "src/public_api.ts"
+    "entryFile": "public_api.ts"
   },
   "whitelistedNonPeerDependencies": [
     "typescript-monads"
