--- projects/flosportsinc/ng-media-source-extensions/src/public_api.ts
@@ -1,7 +1,0 @@
-/*
- * Public API Surface of hls
- */
-
-export * from './lib/mse/mse.module'
-export * from './lib/mse/mse.tokens'
-export * from './lib/mse/mse.directive'
