+++ projects/flosportsinc/ng-media-source-extensions/tsconfig.lib.json
@@ -26,7 +26,7 @@
     "enableResourceInlining": true
   },
   "exclude": [
-    "src/test.ts",
+    "test.ts",
     "**/*.spec.ts"
   ]
 }
