+++ projects/flosportsinc/ng-media-source-extensions/tsconfig.spec.json
@@ -8,9 +8,10 @@
     ]
   },
   "files": [
-    "src/test.ts"
+    "test.ts"
   ],
   "include": [
+    "../*.spec.ts", // your spec file location parent of src or ../yourpath/
     "**/*.spec.ts",
     "**/*.d.ts"
   ]
