+++ projects/flosportsinc/ng-media-source-extensions/tslint.json
@@ -1,17 +1,17 @@
 {
-    "extends": "../../../tslint.json",
-    "rules": {
-        "directive-selector": [
-            true,
-            "attribute",
-            "flo",
-            "camelCase"
-        ],
-        "component-selector": [
-            true,
-            "element",
-            "flo",
-            "kebab-case"
-        ]
-    }
+  "extends": "../../../tslint.json",
+  "rules": {
+    "directive-selector": [
+      true,
+      "attribute",
+      "flo",
+      "camelCase"
+    ],
+    "component-selector": [
+      true,
+      "element",
+      "flo",
+      "kebab-case"
+    ]
+  }
 }
