+++ src/app/app.browser.module.ts
@@ -2,8 +2,8 @@ import { NgModule } from '@angular/core'
 import { PwaService } from './pwa.service'
 import { AppModule } from './app.module'
 import { AppComponent } from './app.component'
-import { DashModule } from '@flosportsinc/ng-media-source-extensions-dash'
-import { HlsModule } from '@flosportsinc/ng-media-source-extensions-hls'
+import { DashModule } from '@flosportsinc/ng-media-source-extensions/dash'
+import { HlsModule } from '@flosportsinc/ng-media-source-extensions/hls'
 import { NodeEnvTransferBrowserModule } from '@flosportsinc/ng-env-transfer-state/browser'
 import { AdBlockBrowserModule } from '@flosportsinc/ng-ad-block/browser'
 import { CookieBrowserModule } from '@flosportsinc/ng-universal-services/cookies/browser'
@@ -13,9 +13,9 @@ import { SvgTransferStateBrowserModule } from '@flosportsinc/ng-svg-transfer-sta
   imports: [
     AdBlockBrowserModule.usingUrl('/assets/ads.js'),
     CookieBrowserModule,
+    AppModule,
     HlsModule,
     DashModule,
-    AppModule,
     NodeEnvTransferBrowserModule.config({
       mergeWithServer: {
         test: 'ok'
