+++ src/app/hls/hls.component.spec.ts
@@ -1,6 +1,6 @@
 import { TestBed, async } from '@angular/core/testing'
 import { HlsComponent } from './hls.component'
-import { HlsModule } from '@flosportsinc/ng-media-source-extensions-hls'
+import { HlsModule } from '@flosportsinc/ng-media-source-extensions/hls'
 import { SharedTestingModule } from '../shared.testing.module'
 
 describe('HlsComponent', () => {
